+++ android/src/main/kotlin/vn/vihat/omicall/omicallsdk/constants/constanst.kt
@@ -36,4 +36,5 @@ const val SPEAKER = "SPEAKER"
 const val VIDEO = "VIDEO"
 const val MUTED = "MUTED"
 const val CLICK_MISSED_CALL = "CLICK_MISSED_CALL"
+const val SWITCHBOARD_ANSWER = "SWITCHBOARD_ANSWER"