+++ ios/Classes/CallProcess/CallManager.swift
@@ -127,6 +127,11 @@ class CallManager {
                                                    name: NSNotification.Name.OMICallDealloc,
                                                    object: nil
             )
+            NotificationCenter.default.addObserver(CallManager.instance!,
+                                                   selector: #selector(self.switchBoardAnswer(_:)),
+                                                   name: NSNotification.Name.OMICallSwitchBoardAnswer,
+                                                   object: nil
+            )
             self.showMissedCall()
         }
     }
@@ -162,6 +167,15 @@ class CallManager {
         DispatchQueue.main.async {
             NotificationCenter.default.removeObserver(CallManager.instance!, name: NSNotification.Name.OMICallVideoInfo, object: nil)
         }
+    }
+    
+    @objc func switchBoardAnswer(_ notification: NSNotification) {
+        guard let userInfo = notification.userInfo,
+              let sip     = userInfo[OMINotificationSIPKey] as? String else {
+            return;
+        }
+        guestPhone = sip
+        SwiftOmikitPlugin.instance?.sendEvent(SWITCHBOARD_ANSWER, ["sip": sip])
     }
     
     @objc func callDealloc(_ notification: NSNotification) {