+++ lib/constant/events.dart
@@ -8,5 +8,6 @@ class OmiEventList {
   static const String onLocalVideoReady = "LOCAL_VIDEO_READY";
   static const String onRemoteVideoReady = "REMOTE_VIDEO_READY";
   static const String onMissedCall = "CLICK_MISSED_CALL";
+  static const String onSwitchboardAnswer = "SWITCHBOARD_ANSWER";
 
 }